usbacc: Move usb driver structs and funcs to usbbus.h and clean-up some usbacc functions
